Add ARK report to all collections
fixes #2338
