+++ app/resources/collections/collections_controller.rb
@@ -1,4 +1,6 @@
 # frozen_string_literal: true
+require "csv"
+
 class CollectionsController < ApplicationController
   include ResourceController
   include TokenAuth
@@ -26,4 +28,37 @@ class CollectionsController < ApplicationController
       end
     end
   end
+
+  def ark_report
+    collection = query_service.find_by(id: params[:id])
+    authorize! :read, collection
+    @collection = collection.decorate
+    @records = ark_report_records(@collection)
+    respond_to do |format|
+      format.html
+      format.csv do
+        send_data ark_csv(@records),
+                  filename: "ark-report-#{@collection.title.parameterize}-#{Time.zone.today}.csv"
+      end
+    end
+  end
+
+  private
+
+    def ark_report_records(collection)
+      collection.members.map do |resource|
+        [resource.source_metadata_identifier&.first,
+         resource.identifier&.first,
+         helpers.manifest_url(resource)]
+      end
+    end
+
+    def ark_csv(records)
+      CSV.generate(headers: true) do |csv|
+        csv << ["source_metadata_id", "ark", "manifest_url"]
+        records.each do |record|
+          csv << record
+        end
+      end
+    end
 end
