--- app/views/archival_media_collections/ark_report.html.erb
@@ -1,21 +1,0 @@
-<h2>ARK / Component ID correlations</h2>
-<p>
-<%= link_to "Download CSV", archival_media_collections_ark_report_path(format: :csv) %>
-</p>
-
-<table class="table table-striped datatable">
-  <thead>
-    <tr>
-      <th>ARK</th>
-      <th>Component ID</th>
-    </tr>
-  </thead>
-  <tbody>
-    <% @resources.each do |r| %>
-      <tr>
-        <td><%= r.identifier.first %></td>
-        <td><%= r.source_metadata_identifier.first %></td>
-      </tr>
-    <% end %>
-  </tbody>
-</table>
