+++ app/views/collections/ark_report.html.erb
@@ -1,0 +1,23 @@
+<h2>ARK Report: <%= @collection.title %></h2>
+<p>
+<%= link_to "Download CSV", collections_ark_report_path(format: :csv) %>
+</p>
+
+<table class="table table-striped datatable">
+  <thead>
+    <tr>
+      <th>Source Metadata ID</th>
+      <th>ARK</th>
+      <th>Manifest URL</th>
+    </tr>
+  </thead>
+  <tbody>
+    <% @records.each do |r| %>
+      <tr>
+        <% r.each do |entry| %>
+          <td><%= entry %></td>
+        <% end %>
+      </tr>
+    <% end %>
+  </tbody>
+</table>
