+++ spec/resources/archival_media_collection/archival_media_collection_feature_spec.rb
@@ -14,7 +14,7 @@ RSpec.feature "Refresh" do
       persister.save(resource: FactoryBot.build(:complete_media_resource, member_of_collection_ids: [collection.id]))
       visit "catalog/#{collection.id}"
       expect(page).to have_link "View all 1 items in this collection"
-      expect(page).to have_link "View ARK / Component ID report"
+      expect(page).to have_link "View ARK report"
       expect(page).to have_link "Edit This Archival Media Collection", href: edit_archival_media_collection_path(collection)
       expect(page).to have_link "Delete This Archival Media Collection", href: archival_media_collection_path(collection)
       expect(page).not_to have_link "File Manager"
