+++ spec/resources/archival_media_collection/archival_media_collections_controller_spec.rb
@@ -52,45 +52,5 @@ RSpec.describe ArchivalMediaCollectionsController, type: :controller do
         end.to enqueue_job(IngestArchivalMediaBagJob)
       end
     end
-
-    describe "show ark_report" do
-      let(:collection) { FactoryBot.create_for_repository(:archival_media_collection, source_metadata_identifier: ["C0652"]) }
-      let(:resource) { FactoryBot.build(:complete_media_resource, title: []) }
-      let(:change_set_persister) { ChangeSetPersister.new(metadata_adapter: Valkyrie.config.metadata_adapter, storage_adapter: Valkyrie.config.storage_adapter) }
-      let(:data) { "ark,component_id\nark:/99999/fk47564298,C0652_c0377\n" }
-
-      before do
-        stub_ezid(shoulder: "99999/fk4", blade: "7564298")
-        stub_pulfa(pulfa_id: "C0652_c0377")
-
-        change_set = MediaResourceChangeSet.new(resource)
-        change_set.validate(source_metadata_identifier: "C0652_c0377", member_of_collection_ids: [collection.id])
-        change_set_persister.save(change_set: change_set)
-      end
-
-      it "displays html" do
-        get :ark_report, params: { id: collection.id }
-
-        expect(response).to render_template :ark_report
-      end
-
-      it "allows downloading a CSV file" do
-        get :ark_report, params: { id: collection.id, format: "csv" }
-
-        expect(response.body).to eq(data)
-      end
-    end
-  end
-
-  context "when an anonymous user" do
-    describe "show ark_report" do
-      let(:collection) { FactoryBot.create_for_repository(:archival_media_collection, source_metadata_identifier: ["C0652"], state: "draft") }
-
-      it "does not display" do
-        get :ark_report, params: { id: collection.id }
-
-        expect(response).to be_redirect
-      end
-    end
   end
 end
